+++ delphin/sembase.py
@@ -10,21 +10,35 @@ def role_priority(role):
     )
 
 
+_COMMON_PROPERTIES = (
+    'PERS',      # [x] person (ERG, Jacy)
+    'NUM',       # [x] number (ERG, Jacy)
+    'GEND',      # [x] gender (ERG, Jacy)
+    'IND',       # [x] individuated (ERG)
+    'PT',        # [x] pronoun-type (ERG)
+    'PRONTYPE',  # [x] pronoun-type (Jacy)
+    'SF',        # [e] sentential-force (ERG)
+    'TENSE',     # [e] tense (ERG, Jacy)
+    'MOOD',      # [e] mood (ERG, Jacy)
+    'PROG',      # [e] progressive (ERG, Jacy)
+    'PERF',      # [e] perfective (ERG, Jacy)
+    'ASPECT',    # [e] other aspect (Jacy)
+    'PASS',      # [e] passive (Jacy)
+    )
+
+_COMMON_PROPERTY_INDEX = dict((p, i) for i, p in enumerate(_COMMON_PROPERTIES))
+
+
 def property_priority(prop):
     """
     Return a representation of property priority for ordering.
 
     Note:
+
        The ordering provided by this function was modeled on the ERG
-       and Jacy grammars and may be irrelevant for others.
+       and Jacy grammars and may be inaccurate for others. Properties
+       not known to this function will be sorted alphabetically.
     """
-    prop = prop.upper()
-    proplist = (
-        'PERS', 'NUM', 'GEND', 'IND', 'PT', 'PRONTYPE',
-        'SF', 'TENSE', 'MOOD', 'PROG', 'PERF', 'ASPECT', 'PASS'
-    )
-    try:
-        return (proplist.index(prop), prop)
-    except ValueError:
-        return (len(proplist), prop)
+    index = _COMMON_PROPERTY_INDEX.get(prop.upper(), len(_COMMON_PROPERTIES))
+    return (index, prop)